renamed component to smartinput
